Refactor: split out MongoDriverResiliencyTest
